cleaning up sql example
